% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_diff}
\alias{tlsh_diff}
\title{Calculate the distance between two TLSH hashes}
\usage{
out <- tlsh_diff(tlsh_a, tlsh_b)
}
\arguments{
\item{a}{TLSH hash string}

\item{b}{TLSH hash string}
}
\value{
integer representing the distance between the two hashes
}
\description{
negative values indicate one of the two hashes is invalid
}
