% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_from_str}
\alias{tlsh_from_str}
\title{Transform a string into a vector of TLSH hashes}
\usage{
out <- tlsh_from_str(string_val_gt_50)
}
\arguments{
\item{s}{a string}
}
\value{
a TLSH hash
}
\description{
Transform a string into a vector of TLSH hashes
}
