% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_from_strs}
\alias{tlsh_from_strs}
\title{Transform a vector of strings into a vector of TLSH hashes}
\usage{
df <- read.csv('test.csv')
names(df) <- c('raw_string_value')
df <- df %>%
  mutate(tlsh_hash = tlsh_from_strs(raw_string_value))
}
\arguments{
\item{xs}{vector of strings}
}
\value{
vector of TLSH hashes
}
\description{
Transform a vector of strings into a vector of TLSH hashes
}
