% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_n_pairs}
\alias{tlsh_n_pairs}
\title{Given a vector of tlsh hashes, return a count of all the pairwise matches
given some threshold}
\usage{
out <- tlsh_n_pairs(10, hashes)
}
\arguments{
\item{threshold}{is the maximum distance to be considered a pair}

\item{xs}{is a vector of TLSH hashes}
}
\value{
a count of how many pairs are in the vector
}
\description{
Given a vector of tlsh hashes, return a count of all the pairwise matches
given some threshold
}
