% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_n_union}
\alias{tlsh_n_union}
\title{Compare two sets of TLSH hashes, return the count of how many
similar hashes there are.}
\usage{
out <- tlsh_n_union(tlsh_set_a, tlsh_set_b)
}
\arguments{
\item{xs}{list of TLSH hashes}

\item{xs2}{list of TLSH hashes}
}
\value{
integer showing the union count between the two sets.
}
\description{
calculates the union of two TLSH lists given some
  threshold.
}
