% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_tree_add}
\alias{tlsh_tree_add}
\title{Adds a TLSH hash to an existing tree.}
\usage{
tree <- tlsh_tree_add(tree, tlsh_hash)
}
\arguments{
\item{key_str}{is the TLSH hash to be added}

\item{tlshtree}{value from creating with tlsh_tree_new}
}
\value{
a pointer to the tlshtree
}
\description{
Adds a TLSH hash to an existing tree. This will not
rebuild the structure of the existing tree, only the hash will be added to
the appropriate leaf. The
underlying point value of the tree will not change, but this function
returns that pointer value regardless, so the return value can be
ignored for most use cases.
}
