% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_tree_delete}
\alias{tlsh_tree_delete}
\title{Deletes a TLSH hash from an existing tree.}
\usage{
tree <- tlsh_tree_delete(tree, tlsh_hash)
}
\arguments{
\item{key_str}{is the TLSH hash to be deleted}

\item{tlshtree}{value from creating with tlsh_tree_new}
}
\value{
a pointer to the tlshtree
}
\description{
Deletes a TLSH hash from an existing tree.
}
