% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_tree_info}
\alias{tlsh_tree_info}
\title{Returns simple information about the tree}
\usage{
out <- tlsh_tree_info(tree)
}
\arguments{
\item{tlshtree}{value from creating with tlsh_tree_new}
}
\value{
List object with information about the tree
}
\description{
Returns simple information about the tree
}
