% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_tree_n_matches}
\alias{tlsh_tree_n_matches}
\title{For each TLSH hash, count how many matches it has within the tree, given
some threshold}
\usage{
out <- tlsh_tree_n_matches(10, hashes)
}
\arguments{
\item{threshold}{is the maximum distance to be considered a pair}

\item{tlshtree}{value from creating with tlsh_tree_new}

\item{xs}{is a vector of TLSH hashes}
}
\value{
a count of how many pairs are in the vector
}
\description{
For each TLSH hash, count how many matches it has within the tree, given
some threshold
}
