% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tlsh_tree_new}
\alias{tlsh_tree_new}
\title{Creates a new TLSH tree given the vector of TLSH hashes}
\usage{
out <- tlsh_tree_new(hashes, 20)
}
\arguments{
\item{leaf_size}{is the set size of each leaf, minimum size is 20}

\item{xs}{is a vector of TLSH hashes}
}
\value{
tlshtree pointer
}
\description{
Creates a new TLSH tree given the vector of TLSH hashes
}
