\name{tlshtree-package}
\alias{tlshtree-package}
\alias{tlshtree}
\docType{package}
\title{
\packageTitle{tlshtree}
}
\description{
\packageDescription{tlshtree}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tlshtree}
\packageIndices{tlshtree}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tlshtree}

Maintainer: \packageMaintainer{tlshtree}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
